+++ packages/tools/generator-widget/generators/app/lib/text.js
@@ -6,42 +6,31 @@ const pkg = require(join(__dirname, "../../../package.json"));
 module.exports = {
     BANNER: `
 
-    ${chalk.bold.bgBlueBright("  __  ____   __ ")}           _     _            _    
-    ${chalk.bold.bgBlueBright(" |  \\/  \\ \\ / / ")}          (_)   | |          | |   
-    ${chalk.bold.bgBlueBright(" | \\  / |\\ V /  ")} __      ___  __| | ____  ___| |_  
-    ${chalk.bold.bgBlueBright(" | |\\/| | > <   ")} \\ \\ /\\ / / |/ _  |/ _  |/ _ \\ __| 
-    ${chalk.bold.bgBlueBright(" | |  | |/ . \\  ")}  \\ V  V /| | (_| | (_| |  __/ |_  
-    ${chalk.bold.bgBlueBright(" |_|  |_/_/ \\_\\ ")}   \\_/\\_/ |_|\\__._|\\__. |\\___|\\__| 
-    ${chalk.bold.bgBlueBright("                ")}                    __/ |          
+    ${chalk.bgBlueBright(chalk.bold.white("  __  ____   __ "))}           _     _            _    
+    ${chalk.bold.bgBlueBright(chalk.bold.white(" |  \\/  \\ \\ / / "))}          (_)   | |          | |   
+    ${chalk.bold.bgBlueBright(chalk.bold.white(" | \\  / |\\ V /  "))} __      ___  __| | ____  ___| |_  
+    ${chalk.bold.bgBlueBright(chalk.bold.white(" | |\\/| | > <   "))} \\ \\ /\\ / / |/ _  |/ _  |/ _ \\ __| 
+    ${chalk.bold.bgBlueBright(chalk.bold.white(" | |  | |/ . \\  "))}  \\ V  V /| | (_| | (_| |  __/ |_  
+    ${chalk.bold.bgBlueBright(chalk.bold.white(" |_|  |_/_/ \\_\\ "))}   \\_/\\_/ |_|\\__._|\\__. |\\___|\\__| 
+    ${chalk.bold.bgBlueBright(chalk.bold.white("                "))}                    __/ |          
                                        |___/           
-     Generator, version: ${pkg.version}
+     Widget Generator, version: ${pkg.version}
                 
     `,
 
-    PACKAGE_READ_ERROR:
-        "Error reading package.json. Please check the file or remove it before you run the generator again. Error: ",
     DIR_NOT_EMPTY_ERROR: chalk.bold.red(
         "The directory you are trying to use is not empty, please open the generator in an empty folder or type yo @mendix/widget WidgetName\n"
     ),
-
     INSTALL_FINISH_MSG:
-        "File configuration done, now running " +
-        chalk.blueBright("npm install") +
-        " to install development dependencies",
+        "File configuration done, now running " + chalk.blueBright("npm install") + " to install dependencies",
     END_NPM_NEED_INSTALL_MSG:
         "\n\n> Dependencies should be installed using " +
         chalk.bold.blueBright("npm install") +
         " before I can run the build using " +
-        chalk.blueBright("gulp") +
+        chalk.blueBright("Pluggable Widgets Tools") +
         " < \n\n",
-    END_RUN_BUILD_MSG_PATH:
-        "\n> Copied files, now running " +
-        chalk.blueBright("npm config set") +
-        " to set your widget path (change this path based on you configurations)) \n\n",
     END_RUN_BUILD_MSG:
-        "\n\n> I will now run " +
-        chalk.bold.blueBright("gulp") +
-        " to build the mpk (do this before open Mendix Studio Pro)< \n\n",
+        "\n\n> I will now run " + chalk.bold.blueBright("Pluggable Widgets Tools") + " to build the widget mpk. < \n\n",
     END_SUCCESS:
         "\n\n\n Widget successfully built!! Please open " +
         chalk.bold.blueBright("Mendix Studio Pro") +