+++ packages/tools/generator-widget/CHANGELOG.md
@@ -6,8 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+- We've added missing functions in editorPreview files.
+- We've added editorConfig file for widgets.
+
 ### Changed
 - We've updated the editorPreview files for web projects with correct properties.
+- We've updated the version of yeoman-generator to 5.4.2.
+- We've updated the configurations for tsconfig in web widgets.
+- We've updated the npm tasks and template classes for web widgets.
 
 ## [9.0.2] - 2021-05-20